requirements.txt
flask==3.1.0
flask-restx==1.3.0
flask-cors==5.0.0
flask-limiter==3.8.0
google-generativeai==0.8.4
transformers==4.48.0
torch==2.5.1
nltk==3.9.1
spacy==3.7.6
python-dotenv==1.0.1
streamlit==1.41.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
requests==2.32.3
